Testnet port value of 119985 is too long for 'unsigned short', changed     to 'unsigned long' to compensate. Probably doesn't matter as testnet     is disabled.  This fixes the warning, though.
        src/protocol.h:21:22: warning: implicit conversion from 'int' to
          'unsigned short' changes value from 119985 to 54449

+++ src/protocol.h
@@ -16,7 +16,7 @@
 #include "uint256.h"
 
 extern bool fTestNet;
-static inline unsigned short GetDefaultPort(const bool testnet = fTestNet)
+static inline unsigned long GetDefaultPort(const bool testnet = fTestNet)
 {
     return testnet ? 119985 : 19985;
 }
